Working but still needs TLC with automatically providing patch

+++ public/Install-KbUpdate.ps1
@@ -21,6 +21,9 @@ function Install-KbUpdate {
     .PARAMETER FilePath
         The filepath of the patch. This needs to be updated to be more in-depth.
 
+    .PARAMETER Guid
+        The filepath of the patch. This needs to be updated to be more in-depth.
+
     .PARAMETER Type
         The type of patch. Basically General or SQL.
 
@@ -39,16 +42,30 @@ function Install-KbUpdate {
         PS C:\> Install-KbUpdate -ComputerName sql2017 -HotfixId KB4534273 -FilePath C:\temp\windows10.0-kb4534273-x64_74bf76bc5a941bbbd0052caf5c3f956867e1de38.msu
 
         Installs KB4534273 from the C:\temp directory on sql2017
+
+    .EXAMPLE
+        PS C:\> Get-KbUpdate -Pattern 4498951 | Install-KbUpdate -Type SQL -ComputerName sql2017
+
+        Installs KB4534273 from the C:\temp directory on sql2017
 #>
-    [CmdletBinding(SupportsShouldProcess, ConfirmImpact = "High")]
+    [CmdletBinding(SupportsShouldProcess, ConfirmImpact = "Medium")]
     param (
         [string[]]$ComputerName = $env:ComputerName,
         [PSCredential]$Credential,
         [PSCredential]$PSDscRunAsCredential,
         [string]$HotfixId,
         [string]$FilePath,
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [Alias("UpdateId")]
+        [string]$Guid,
         [ValidateSet("General", "SQL")]
         [string]$Type = "General",
+        [Parameter(ValueFromPipelineByPropertyName)]
+        [string]$Title,
+        [switch]$NoDelete,
+        [Parameter(ValueFromPipeline)]
+        [pscustomobject]$InputObject,
+        [switch]$Force,
         [switch]$EnableException
     )
     begin {
@@ -57,28 +74,31 @@ function Install-KbUpdate {
         }
     }
     process {
-        if (-not $PSBoundParameters.HotfixId -and -not $PSBoundParameters.FilePath) {
-            Stop-Function -EnableException:$EnableException -Message "You must specify either HotfixId or FilePath"
+        if (-not $PSBoundParameters.HotfixId -and -not $PSBoundParameters.FilePath -and -not $PSBoundParameters.InputObject) {
+            Stop-Function -EnableException:$EnableException -Message "You must specify either HotfixId or FilePath or pipe in the results from Get-KbUpdate"
             return
         }
 
         foreach ($computer in $ComputerName) {
-            $exists = Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ArgumentList $HotfixId -ScriptBlock {
-                Get-HotFix -Id $args -ErrorAction SilentlyContinue
+            #null out a couple things to be safe
+            $remoteexists = $remotehome = $remotesession = $null
+            if ($HotFixId) {
+                $exists = Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ArgumentList $HotfixId -ScriptBlock {
+                    Get-HotFix -Id $args -ErrorAction SilentlyContinue
+                }
             }
 
             if ($exists) {
                 Stop-Function -EnableException:$EnableException -Message "$hotfixid is already installed on $computer" -Continue
             }
 
+            # a lot of the file copy work will be done in the remote $home dir
+            $remotehome = Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ScriptBlock { $home }
             $remotesession = Get-PSSession -ComputerName $computer | Where-Object Availability -eq Available
 
             if (-not $remotesession) {
                 Stop-Function -EnableException:$EnableException -Message "Session for $computer can't be found or no runspaces are available. Please file an issue on the GitHub repo at https://github.com/potatoqualitee/kbupdate/issues" -Continue
             }
-
-            # a lot of the file copy work will be done in the remote $home dir
-            $remotehome = Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ScriptBlock { $home }
 
             # Copy every time even if remote computer has xWindowsUpdate because this version has Jess' fix
             $null = Copy-Item -Path "$script:ModuleRoot\library\xWindowsUpdate" -Destination "$remotehome\kbupdatetemp\xWindowsUpdate" -ToSession $remotesession -Recurse -Force
@@ -88,26 +108,45 @@ function Install-KbUpdate {
                 Get-Module -Listavailable -Name SqlServerDsc
             }
             if (-not $sqldscexists -and $Type -eq "SQL") {
-                $null = Copy-Item -Path "$script:ModuleRoot\library\sqlserverdsc" -Destination "$remotehome\kbupdatetemp\sqlserverdsc" -ToSession $remotesession -Recurse -Force
+                #$null = Copy-Item -Path "$script:ModuleRoot\library\sqlserverdsc" -Destination "$remotehome\kbupdatetemp\sqlserverdsc" -ToSession $remotesession -Recurse -Force
             }
 
             if ($PSBoundParameters.FilePath) {
-                $remoteexists = Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ArgumentList $FileName -ScriptBlock { Get-ChildItem -Path $args }
+                $remoteexists = Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ArgumentList $FilePath -ScriptBlock { Get-ChildItem -Path $args -ErrorAction SilentlyContinue }
             }
 
-            if ((-not $remoteexists -or -not $PSBoundParameters.FilePath)) {
-                if ($PSCmdlet.ShouldProcess($computer, "File not detected, downloading now and copying to remote computer")) {
-                    if (-not $updatefile) {
-                        $updatefile = Get-KbUpdate -ComputerName $computer -Architecture x64 -Credential $credential -Latest -Pattern $HotfixId | Select-Object -First 1 | Save-KbUpdate -Path $home
+            if (-not $remoteexists) {
+                if (-not $updatefile) {
+                    # need to detect being piped in via InputObject
+                    if (-not $PSBoundParameters.InputObject) {
+                        $InputObject = Get-KbUpdate -ComputerName $computer -Architecture x64 -Credential $credential -Latest -Pattern $HotfixId
                     }
-                    if (-not $FilePath) {
-                        $FilePath = "$remotehome\$(Split-Path -Leaf $updateFile)"
-                    }
-                    if ($updatefile) {
-                        $null = Copy-Item -Path $updatefile -Destination $FilePath -ToSession $remotesession -Force
+
+                    $file = Split-Path $InputObject.Link -Leaf
+
+                    if ((Test-Path -Path "$home\$file")) {
+                        $updatefile = Get-ChildItem -Path "$home\$file"
                     } else {
-                        Stop-Function -EnableException:$EnableException -Message "Could not find $HotfixId and no file was specified" -Continue
+                        if ($PSCmdlet.ShouldProcess($computer, "File not detected, downloading now and copying to remote computer")) {
+                            $updatefile = $InputObject | Select-Object -First 1 | Save-KbUpdate -Path $home
+                        }
                     }
+                }
+
+                if (-not $FilePath) {
+                    $FilePath = "$remotehome\kbupdatetemp\$(Split-Path -Leaf $updateFile)"
+                }
+
+                if ($updatefile) {
+                    try {
+                        $null = Copy-Item -Path $updatefile -Destination $FilePath -ToSession $remotesession -ErrrorAction Stop
+                    } catch {
+                        $exists = Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ArgumentList $FilePath -ScriptBlock {
+                            Remove-Item $args -Force -ErrorAction SilentlyContinue
+                        }
+                    }
+                } else {
+                    Stop-Function -EnableException:$EnableException -Message "Could not find $HotfixId and no file was specified" -Continue
                 }
             }
 
@@ -116,7 +155,8 @@ function Install-KbUpdate {
                 #windows10.0-kb4516115-x64
                 $HotfixId = $FilePath.ToUpper() -split "\-" | Where-Object { $psitem.Startswith("KB") }
                 if (-not $HotfixId) {
-                    Stop-Function -EnableException:$EnableException -Message "Could not determine KB from $FilePath. Looked for '-kbnumber-'" -Continue
+                    Stop-Function -EnableException:$EnableException -Message "Could not determine KB from $FilePath. Looked for '-kbnumber-'. Please provide a HotfixId."
+                    return
                 }
             }
 
@@ -146,43 +186,70 @@ function Install-KbUpdate {
                     }
                 }
                 "SQL" {
-                    # dont know yet
+                    $hotfix = @{
+                        Name       = 'Package'
+                        ModuleName = 'PSDesiredStateConfiguration'
+                        Property   = @{
+                            Ensure     = 'Present'
+                            Name       = $Title
+                            ProductId  = $Guid
+                            Path       = $FilePath
+                            Arguments  = "/action=patch /AllInstances /quiet /IAcceptSQLServerLicenseTerms"
+                            ReturnCode = 0, 3010
+                        }
+                    }
                 }
             }
 
             ## could also use xPendingReboot to look for pending reboots and handle?
+
             if ($PSCmdlet.ShouldProcess($computer, "Installing Hotfix $HotfixId from $FilePath")) {
-                Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ScriptBlock {
-                    param (
-                        $Hotfix,
-                        $Verbose
-                    )
-                    if (-not (Get-Command Invoke-DscResource)) {
-                        throw "Invoke-DscResource not found on $env:ComputerName"
-                    }
-                    # Extract exes, cabs? exe = /extract
-                    Import-Module "$home\xWindowsUpdate" -Force
-                    Write-Verbose -Message ("Installing {0} from {1}" -f $hotfix.property.id, $hotfix.property.path)
-                    try {
-                        if (-not (Invoke-DscResource @hotfix -Method Test -Verbose)) {
-                            Invoke-DscResource @hotfix -Method Set -Verbose
+                try {
+                    Invoke-PSFCommand -ComputerName $computer -Credential $Credential -ScriptBlock {
+                        param (
+                            $Hotfix,
+                            $VerbosePreference,
+                            $NoDelete,
+                            $ManualFileName
+                        )
+
+                        if (-not (Get-Command Invoke-DscResource)) {
+                            throw "Invoke-DscResource not found on $env:ComputerName"
                         }
-                    } catch {
-                        # sometimes there's a "Serialized XML" issue that can be ignored because
-                        # the patch installs successfully anyway. so throw only if there's a real issue
-                        # Serialized XML is nested too deeply. Line 1, position 3507.
-                        if ($_.Exception.Message -notmatch "Serialized XML is nested too deeply") {
-                            throw
+                        # Extract exes, cabs? exe = /extract
+                        Import-Module "$home\kbupdatetemp\xWindowsUpdate" -Force
+                        Write-Verbose -Message ("Installing {0} from {1}" -f $hotfix.property.id, $hotfix.property.path)
+                        try {
+                            if (-not (Invoke-DscResource @hotfix -Method Test)) {
+                                Invoke-DscResource @hotfix -Method Set -ErrorAction Stop
+                            }
+                        } catch {
+                            # sometimes there's a "Serialized XML" issue that can be ignored because
+                            # the patch installs successfully anyway. so throw only if there's a real issue
+                            # Serialized XML is nested too deeply. Line 1, position 3507.
+                            if ("$_" -match "Serialized XML is nested too deeply") {
+                                continue
+                            } elseif ("$_" -match "2042429437") {
+                                throw "The return code -2042429437 was not expected. Configuration is likely not correct. The requested features may not be installed or features are already at a higher patch level."
+                            } else {
+                                throw
+                            }
                         }
-                    }
-                    Remove-Module SqlServerDsc -ErrorAction SilentlyContinue
-                    Remove-Module xWindowsUpdate -ErrorAction SilentlyContinue
-                    Remove-Item -Recurse -Force -ErrorAction SilentlyContinue -Path "$home\kbupdatetemp"
-                } -ArgumentList $hotfix, $PSBoundParameters.Verbose
+                        Remove-Module SqlServerDsc -ErrorAction SilentlyContinue
+                        Remove-Module xWindowsUpdate -ErrorAction SilentlyContinue
+                        if (-not $NoDelete -and -not $ManualFileName) {
+                            Remove-Item -Recurse -Force -ErrorAction SilentlyContinue -Path "$home\kbupdatetemp"
+                        }
+                    } -ArgumentList $hotfix, $VerbosePreference, $NoDelete, $PSBoundParameters.FileName -ErrorAction Stop
+                } catch {
+                    Stop-Function -Message "Failure on $computer" -ErrorRecord $_ -EnableException:$EnableException
+                }
             }
         }
-
-        if ($updatefile) {
+    }
+    end {
+        # if copy-item tosession remote fails, we leave litter. gotta fix that.
+        if ($updatefile -and -not $NoDelete) {
             Remove-Item -Path $Path -ErrorAction SilentlyContinue
         }
     }
